Refactor stations list module (#333)
