+++ src/Configuration/Redux/Reducers.ts
@@ -1,6 +1,6 @@
+import { IStationsListState, stationsReducer } from 'Modules/Station/Redux';
 import { chatReducer, playlistReducer } from 'Modules/Station/Redux/Reducer';
 import { IMessage, IPlaylistState } from 'Modules/Station/Redux/Types';
-import { IStationsListState, stationsReducer } from 'Modules/Stations/Redux';
 import { favoriteReducer, userReducer } from 'Modules/User/Redux/Reducer';
 import { IUser } from 'Modules/User/Redux/Types';
 import { IFavorite } from 'Modules/User/Redux/Types/Types';
