+++ src/Modules/Station/Redux/Constants/StationsList.ts
@@ -1,4 +1,4 @@
-export const actionTypes = {
+export const stationsListActionTypes = {
   STATIONS_LIST_UPDATE: 'STATIONS_BROWSER:LIST_UPDATED',
   STATIONS_LIST_CONNECTED: 'STATIONS_BROWSER:LIST_CONNECTED',
   STATIONS_LIST_ERROR: 'STATIONS_BROWSER:LIST_ERROR',
