+++ src/Modules/Station/Redux/Types/Types.ts
@@ -1,3 +1,4 @@
 export * from './IPlaylist';
 export * from './ISong';
 export * from './IMessage';
+export * from './StationsList';
