--- src/Modules/Stations/Redux/index.ts
@@ -1,3 +1,0 @@
-export * from './Reducers';
-export * from './Types';
-export * from './Constants';
