--- src/Modules/Stations/index.ts
@@ -1,1 +1,0 @@
-export * from './Redux';
